Remove graph + move
