+++ src/metacity/geometry/mesh/model.cpp
