--- src/metacity/geometry/vector_pipeline/feature.hpp
@@ -1,4 +1,0 @@
-
-
-
-
