+++ src/metacity/io/__init__.py
@@ -1,6 +1,5 @@
 from metacity.io.geojson import parse as parse_geojson
 from metacity.io.shapefile import parse as parse_shapefile
-from metacity.io.osm_graph import parse_graph as parse_osm_graph
 import metacity.utils.filesystem as fs
 from metacity.geometry import Progress 
 
@@ -16,9 +15,6 @@ def parse(file: str, progress: Progress = None):
     elif file.endswith('.json'):
         return parse_geojson(file, progress)
 
-
-def parse_graph(node_file: str, edge_file: str):
-    return parse_osm_graph(node_file, edge_file)
 
 
 def parse_recursively(directory: str):
